# api-zod__generated_types_orderInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * E-commerce store API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderInputItemsItem } from './orderInputItemsItem';

export interface OrderInput {
  customerId?: number;
  customerName: string;
  customerEmail: string;
  customerPhone?: string;
  shippingAddress: string;
  shippingCity: string;
  shippingCountry: string;
  notes?: string;
  sessionId?: string;
  items: OrderInputItemsItem[];
}