# api-zod__generated_types_ordersPage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * E-commerce store API
 * OpenAPI spec version: 0.1.0
 */
import type { Order } from './order';

export interface OrdersPage {
  orders: Order[];
  total: number;
  page: number;
  limit: number;
}