# api-zod__generated_types_product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * E-commerce store API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  comparePrice?: number | null;
  /** @nullable */
  imageUrl?: string | null;
  images?: string[];
  stock: number;
  /** @nullable */
  sku?: string | null;
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  isFeatured?: boolean;
  /** @nullable */
  rating?: number | null;
  reviewCount?: number;
  createdAt?: string;
}