# api-zod__generated_types_productsPage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * E-commerce store API
 * OpenAPI spec version: 0.1.0
 */
import type { Product } from './product';

export interface ProductsPage {
  products: Product[];
  total: number;
  page: number;
  limit: number;
}